+++ Sources/MapboxMaps/Foundation/Extensions/UIViewController+Extensions.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+extension UIViewController {
+    var topmostPresentedViewController: UIViewController {
+        var result = self
+        while let presented = result.presentedViewController {
+            result = presented
+        }
+        return result
+    }
+}