+++ Tests/MapboxMapsTests/Style/Mocks/MockAttributionDialogManagerDelegate.swift
@@ -5,7 +5,7 @@ import Foundation
 final class MockAttributionDialogManagerDelegate: AttributionDialogManagerDelegate {
 
     let viewControllerForPresentingStub = Stub<AttributionDialogManager, UIViewController>(defaultReturnValue: UIViewController())
-    func viewControllerForPresenting(_ attributionDialogManager: AttributionDialogManager) -> UIViewController {
+    func viewControllerForPresenting(_ attributionDialogManager: AttributionDialogManager) -> UIViewController? {
         viewControllerForPresentingStub.call(with: attributionDialogManager)
     }